// client/src/constants/socketEvents.tsx
enum SocketEvents {
  NEW_MESSAGE = "NEW_MESSAGE",
  // game
  CREATE_NEW_GAME = "CREATE_NEW_GAME",
  CREATED_NEW_GAME = "CREATED_NEW_GAME",
  JOIN_GAME = "JOIN_GAME",
  EXIT_FROM_GAME = "EXIT_FROM_GAME",
  UPDATE_BOARD = "UPDATE_BOARD",
  // list
  GET_GAMES_LIST = "GET_GAMES_LIST",
  RECEIVED_GAMES_LIST = "RECEIVED_GAMES_LIST",
  REFRESH_GAMES_LIST = "REFRESH_GAMES_LIST",
  // player

  PLAYER_MOVE = "PLAYER_MOVE"
}

export default SocketEvents;